Improve portability of tests.
Certain assets and libraries were assumed to lie outside the repo in fixed relative locations. As this cannot be guaranteed, I've imported current versions of the Kiwi and Kiwifroot libraries into the repo.

I've also modified `tests/playground.html` to generate a more accurate relative path to the game on test launch.